requirements.txt
scikit-learn-0.19.0
tensorflow-2.4.1
keras==2.4.3
pandas=1.1.5
